requirements.txt
flask>=3.1.0
flask-login>=0.6.3
flask-sqlalchemy>=3.1.0
flask-bcrypt>=1.0.1
flask-limiter>=3.0
python-dotenv>=1.0.0
google-generativeai>=0.8.0
SQLAlchemy>=2.0.0
Werkzeug>=3.0.0
